#5211 - AI assistant prototype
- Try fixing issue when trying to delete a project which is still being reindexed / reindexing a project which is being deleted
