+++ inception/inception-search-core/src/main/java/de/tudarmstadt/ukp/inception/search/cli/SearchReindexCliCommand.java
@@ -33,8 +33,8 @@ import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.inception.project.api.ProjectService;
 import de.tudarmstadt.ukp.inception.scheduling.LoggingTaskMonitor;
 import de.tudarmstadt.ukp.inception.scheduling.SchedulingService;
+import de.tudarmstadt.ukp.inception.search.ReindexTask;
 import de.tudarmstadt.ukp.inception.search.SearchService;
-import de.tudarmstadt.ukp.inception.search.scheduling.tasks.ReindexTask;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
