+++ inception/inception-search-core/src/main/java/de/tudarmstadt/ukp/inception/search/config/SearchServiceAutoConfiguration.java
@@ -30,7 +30,6 @@ import de.tudarmstadt.ukp.inception.documents.api.DocumentService;
 import de.tudarmstadt.ukp.inception.preferences.PreferencesService;
 import de.tudarmstadt.ukp.inception.project.api.ProjectService;
 import de.tudarmstadt.ukp.inception.scheduling.SchedulingService;
-import de.tudarmstadt.ukp.inception.schema.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.inception.schema.api.feature.FeatureSupportRegistry;
 import de.tudarmstadt.ukp.inception.search.FeatureIndexingSupport;
 import de.tudarmstadt.ukp.inception.search.FeatureIndexingSupportRegistry;
@@ -50,12 +49,12 @@ public class SearchServiceAutoConfiguration
 {
     @Bean
     public SearchService searchService(DocumentService aDocumentService,
-            AnnotationSchemaService aSchemaService, ProjectService aProjectService,
-            PhysicalIndexRegistry aPhysicalIndexRegistry, SchedulingService aSchedulingService,
-            SearchServiceProperties aProperties, PreferencesService aPreferencesService)
+            ProjectService aProjectService, PhysicalIndexRegistry aPhysicalIndexRegistry,
+            SchedulingService aSchedulingService, SearchServiceProperties aProperties,
+            PreferencesService aPreferencesService)
     {
-        return new SearchServiceImpl(aDocumentService, aSchemaService, aProjectService,
-                aPhysicalIndexRegistry, aSchedulingService, aProperties, aPreferencesService);
+        return new SearchServiceImpl(aDocumentService, aProjectService, aPhysicalIndexRegistry,
+                aSchedulingService, aProperties, aPreferencesService);
     }
 
     @Bean
